refactor(@sounisi5011/encrypted-archive): stop using Node.js builtin module `events` in core code
